Crystal emitter: emit a standalone typed MethodDef
Adds pub fn emit_method(&MethodDef) -> String to src/emit/crystal.rs:
renders `def [self.]name(p : T, ...) : R ... end` using the attached
Ty::Fn signature. Respects MethodReceiver::Class for `def self.`
(Crystal's module-utility idiom).

Also changes the Ruby source of truth to `def self.pluralize`
(semantically a module function, not an instance method) — Python/TS
emitters ignore receiver so those stay green; Crystal picks up the
`self.` prefix naturally.

Two dogfooding fixes for Crystal:
- emit_expr(If) now emits the ternary form; block form was broken in
  expression position (no existing Rails fixture exercised it).
- emit_send recognizes Ruby binary operators as Crystal infix ops.

Runtime alignment: runtime/crystal/view_helpers.cr:181 `Int` → `Int64`
to match the rest of the Crystal runtime (concrete types throughout).
The emitter is now the source of truth for this function.

Integration test generalized: compares line-by-line modulo leading
whitespace so target files that nest functions inside modules match
cleanly. Adds an assertion for Crystal alongside Python.

All 340 unit+integration tests pass; 3/3 crystal_toolchain tests pass
(including real_blog_crystal_spec_passes, which exercises pluralize
via real-blog views).

+++ runtime/crystal/view_helpers.cr
@@ -178,7 +178,7 @@ module Roundhouse
       prefix.empty? ? base : "#{prefix}_#{base}"
     end
 
-    def self.pluralize(count : Int, word : String) : String
+    def self.pluralize(count : Int64, word : String) : String
       count == 1 ? "1 #{word}" : "#{count} #{word}s"
     end
